requirements.txt
Django==5.1
mysqlclient
djangorestframework==3.15.2
django-cors-headers
django-environ
faker
djangorestframework-simplejwt==5.3.1 # JWT token implementation for DRF
stripe>=11.0.0
